Change check_kernel so it doesn't check the kernel for RELEASE builds update version number to 2.21 cos Debian numbers have to increment like that. add suggests kernel-image-2.4

+++ dnprogs/scripts/check_kernel.sh
@@ -3,6 +3,13 @@
 # Check the kernel headers available to us. 
 #
 #
+
+# But not for production builds. we assume a 2.4 kernel
+if [ -n "$RELEASE" ]
+then
+  return 0
+fi
+
 rm -f include/netdnet/dn.h
 
 if [ -f /usr/src/linux/include/netdnet/dn.h ]
